Ex3&4
Example 3: Construct a function with a local variable. Show what happens when you try to use that variable outside the function. Explain the results.

Example 4: Construct a function that takes an argument. Give the function parameter a unique name. Show what happens when you try to use that parameter name outside the function. Explain the results.