Unit2LearningJournal
Part 1
The volume of a sphere is 4/3πr3, where π has the value of "pi" given in Section 2.1 of your textbook. Write a function called print_volume (r) that takes an argument for the sphere's radius and prints the sphere's volume.

Call your print_volume function three times with different values for radius.